Updated setup.sh

+++ setup.sh
@@ -1,16 +1,21 @@
 #!/bin/bash
+if [ -z "$1" ]; then
+    INSTALL_DIR="/var/tmp"
+else
+	INSTALL_DIR="${1%/}"
+fi
 
 MAIN_DIR=$(pwd)
 
 NAME="mgba-0.5.2"
 FILE="$NAME.tar.gz"
 BUILD_DIR="$NAME"
-INSTALL_DIR="mGBA"
+DIR="mGBA"
 URL="https://github.com/mgba-emu/mgba/archive/0.5.2.tar.gz"
 
-echo "############# mGBA ############"
-if [ ! -d "$INSTALL_DIR" ]; then
-	cd /var/tmp
+echo -e "\e[1m\e[32m############# mGBA ############\e[0m"
+if [ ! -d "$DIR" ]; then
+	cd $INSTALL_DIR
 	if [ ! -f "$FILE" ]; then
 		echo "Downloading..."
 		wget -O $FILE $URL
@@ -42,16 +47,18 @@ if [ ! -d "$INSTALL_DIR" ]; then
 else
 	echo "Already installed"
 fi
+echo -e "\e[1m\e[32m###############################\e[0m"
+echo ""
 
 NAME="devkitARM"
 FILE="$NAME.tar.bz2"
 BUILD_DIR="$NAME"
-INSTALL_DIR="$NAME"
+DIR="$NAME"
 URL="https://sourceforge.net/projects/devkitpro/files/devkitARM/devkitARM_r45/devkitARM_r45-i686-linux.tar.bz2/download"
 
-echo "########## devkitArm ##########"
-if [ ! -d "$INSTALL_DIR" ]; then
-	cd /var/tmp
+echo -e "\e[1m\e[32m########## devkitArm ##########\e[0m"
+if [ ! -d "$DIR" ]; then
+	cd $INSTALL_DIR
 	if [ ! -f "$FILE" ]; then
 		echo "Downloading..."
 		wget -O $FILE $URL
@@ -70,14 +77,16 @@ if [ ! -d "$INSTALL_DIR" ]; then
 else
 	echo "Already installed"
 fi
+echo -e "\e[1m\e[32m###############################\e[0m"
+echo ""
 
 NAME="tiled"
 BUILD_DIR="$NAME"
 URL="https://github.com/JFonS/tiled"
 
-echo "############ Tiled ############"
+echo -e "\e[1m\e[32m############ Tiled ############\e[0m"
 if [ ! -d "$BUILD_DIR" ]; then
-	cd /var/tmp
+	cd $INSTALL_DIR
 	if [ ! -d "$BUILD_DIR" ]; then
 		echo "Downloading..."
 		git clone $URL
@@ -94,27 +103,43 @@ else
 	echo "Already installed"
 fi
 
-echo "###############################"
-echo "Installation done. Please configure your environment."
-echo -e "\e[1m\e[31mThis is your first test and it will be evaluated...\e[0m"
-
-echo ""
+echo -e "\e[1m\e[32m###############################\e[0m"
 echo ""
 
-echo -n "$MAIN_DIR>"
-sleep 15
+if [ ! -f ".lock" ]; then
+	echo "Installation done. Please configure your environment."
+	echo -e "\e[1m\e[31mThis is your first test and it will be evaluated...\e[0m"
 
-echo ""
-echo ""
+	echo ""
+	echo ""
 
-echo -e "\e[1m\e[5m\e[31mCome on... I'm waiting... \e[0m"
-echo ""
-echo ""
+	echo -n "$MAIN_DIR>"
+	sleep 15
 
-sleep 5
-echo -e "\e[1m\e[31mNothing? ok.\e[0m"
-sleep 1
+	echo ""
+	echo ""
 
-echo -e "\e[1m\e[31mIf you are lazy you can just copy the following command:"
-echo -e "\e[1m\e[32msetenv DEVKITARM /var/tmp/devkitARM/ && set path = ( \$path /var/tmp/mgba-0.5.2/build/qt/ /var/tmp/tiled/bin/)\e[0m"
+	echo -e "\e[1m\e[5m\e[31mCome on... I'm waiting... \e[0m"
+	echo ""
+	echo ""
 
+	sleep 5
+	echo -e "\e[1m\e[31mNothing? ok.\e[0m"
+	sleep 1
+
+	echo -e "\e[1m\e[31mIf you are lazy you can just copy the following command:"
+	echo -e "\e[1m\e[32msetenv DEVKITARM $INSTALL_DIR/devkitARM/ && set path = ( \$path $INSTALL_DIR/mgba-0.5.2/build/qt/ $INSTALL_DIR/tiled/bin/)\e[0m"
+	echo ""
+	echo ""
+
+	echo "You might want to add one of the following lines to your init file (.bashrc or .tcshrc):"
+	echo -e "\e[1m\e[31mbash: \e[32mexport DEVKITARM=$INSTALL_DIR/devkitARM/ && export PATH=\${PATH}:$INSTALL_DIR/mgba-0.5.2/build/qt/:$INSTALL_DIR/tiled/bin/\e[0m"
+	echo -e "\e[1m\e[31mtcsh: \e[32msetenv DEVKITARM $INSTALL_DIR/devkitARM/ && set path = (\$path $INSTALL_DIR/mgba-0.5.2/build/qt/ $INSTALL_DIR/tiled/bin/)\e[0m"
+
+	touch ".lock"
+else
+	echo "Installation done. Please configure your environment."
+	echo "Add one of the following lines to your init file (.bashrc or .tcshrc):"
+	echo -e "\e[1m\e[31mbash: \e[32mexport DEVKITARM=$INSTALL_DIR/devkitARM/ && export PATH=\${PATH}:$INSTALL_DIR/mgba-0.5.2/build/qt/:$INSTALL_DIR/tiled/bin/\e[0m"
+	echo -e "\e[1m\e[31mtcsh: \e[32msetenv DEVKITARM $INSTALL_DIR/devkitARM/ && set path = (\$path $INSTALL_DIR/mgba-0.5.2/build/qt/ $INSTALL_DIR/tiled/bin/)\e[0m"
+fi
